Update CVE-2025-22376 following independent review
I did an independent review of this CVE as it felt like the CVSS score was a bit high for what the vulnerability is. I've published the full review on my blog which explains the rationale behind this PR and adjusted CVSS score. I've also added the blog post as a reference.